Updated bill amount format

+++ app/src/main/java/com/frisky/icebreaker/core/structures/OrderSummary.java
@@ -9,7 +9,7 @@ public class OrderSummary {
     private String sessionID;
     private String endTime;
 
-    private int totalAmount;
+    private double totalAmount;
 
     public OrderSummary(String restaurantID, Uri restaurantImage, String restaurantName, String sessionID,
                         String endTime, int totalAmount) {
@@ -41,7 +41,7 @@ public class OrderSummary {
         return endTime;
     }
 
-    public int getTotalAmount() {
+    public double getTotalAmount() {
         return totalAmount;
     }
 }
